+++ ably/event_emitter_spec_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	"github.com/ably/ably-go/ably"
-	"github.com/ably/ably-go/ably/ablytest"
+	"github.com/ably/ably-go/ably/internal/ablytest"
 )
 
 func Test_RTE3_EventEmitter_On(t *testing.T) {
