+++ ably/event_emitter_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/ably/ably-go/ably"
-	"github.com/ably/ably-go/ably/ablytest"
+	"github.com/ably/ably-go/ably/internal/ablytest"
 )
 
 func TestEventEmitterConcurrency(t *testing.T) {
