+++ ably/http_paginated_response_test.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/ably/ably-go/ably"
 
-	"github.com/ably/ably-go/ably/ablytest"
+	"github.com/ably/ably-go/ably/internal/ablytest"
 	"github.com/ably/ably-go/ably/proto"
 )
