+++ ably/proto/message_test.go
@@ -6,7 +6,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/ably/ably-go/ably/ablytest"
+	"github.com/ably/ably-go/ably/internal/ablytest"
 	"github.com/ably/ably-go/ably/proto"
 )
