+++ ably/realtime_presence_test.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/ably/ably-go/ably"
-	"github.com/ably/ably-go/ably/ablytest"
+	"github.com/ably/ably-go/ably/internal/ablytest"
 )
 
 func contains(members []*ably.PresenceMessage, clients ...string) error {
