+++ ably/rest_channel_spec_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	"github.com/ably/ably-go/ably"
-	"github.com/ably/ably-go/ably/ablytest"
+	"github.com/ably/ably-go/ably/internal/ablytest"
 )
 
 func TestRSL1f1(t *testing.T) {
