+++ ably/rest_channel_test.go
@@ -18,7 +18,7 @@ import (
 	"github.com/ably/ably-go/ably/internal/ablyutil"
 
 	"github.com/ably/ably-go/ably"
-	"github.com/ably/ably-go/ably/ablytest"
+	"github.com/ably/ably-go/ably/internal/ablytest"
 	"github.com/ably/ably-go/ably/proto"
 )
